Use explicit scripted upsert for machine ads
This allows both machine and job updates to update the occupancy metric. Also,
the @timestamp field is now a list of every time the machine ad was updated.
This allows the doc to match multiple bins in a date histogram, making a kind
of hobo range indexing possible without upgrading to ES 7.